// types.ts
export interface VTAnalysisStats {
  harmless: number;
  malicious: number;
  suspicious: number;
  undetected: number;
  timeout: number;
}

export interface VTOwner {
    asn?: number;
    as_owner?: string;
    country?: string;
}

export interface VTPcapIdpsRule {
    alert_severity: string;
    alert_context: string[];
    rule_msg: string;
    rule_sid: string;
    rule_gid: string;
    rule_category: string;
}

export interface VTPcapInfo {
    ids_rules?: VTPcapIdpsRule[];
    http_requests?: any[];
    dns_lookups?: any[];
}
// ...
export interface VTFileReport {
  type: 'file';
  id: string; // The hash
  attributes: {
    last_analysis_stats: VTAnalysisStats;
    meaningful_name?: string;
    names?: string[];
    size: number;
    type_description?: string;
    magic?: string;
    last_analysis_date?: number; // timestamp
    creation_date?: number; // timestamp
    signature_info?: { description: string; verified: string; };
    pcap_info?: VTPcapInfo;
  };
  error?: string;
}
// ...
export interface VTDomain {
    id: string;
    type: 'domain';
    attributes: { last_analysis_stats: VTAnalysisStats; reputation: number; };
}

export interface VTFile {
  id: string; // SHA-256 hash
  type: 'file';
  attributes: {
    last_analysis_stats: VTAnalysisStats;
    last_analysis_date: number;
    meaningful_name?: string;
    names?: string[];
  };
}

export interface VTIPAddress {
    id: string;
    type: 'ip_address';
    attributes: { last_analysis_stats: VTAnalysisStats; reputation: number; };
}

export interface VTResolution {
  id: string; // Composite key
  type: 'resolution';
  attributes: {
    date: number; // timestamp
    ip_address: string;
    host_name: string;
  };
}

// Union type for any kind of relationship object VirusTotal might return.
export type VTRelationship = VTFile | VTResolution | VTDomain | VTIPAddress;